requirements.txt
editdistance==0.6.2
fairseq==0.12.2
faiss_cpu==1.7.2
hydra-core==1.3.2
jsonlines==3.1.0
mpi4py==3.1.4
numpy==1.23.1
omegaconf==2.2.2
pytext==1.0
regex==2022.10.31
sentence_transformers==2.2.2
spacy==2.1.8
transformers==3.0.2
wget==3.2
tqdm